refactor: extract Textual-free CLI entry point
