+++ src/kon/cli.py
@@ -1,0 +1,68 @@
+import argparse
+
+from kon import config
+
+from .llm import PROVIDER_API_BY_NAME
+from .version import VERSION
+
+
+def build_parser() -> argparse.ArgumentParser:
+    parser = argparse.ArgumentParser(description="Kon TUI")
+    parser.add_argument("--model", "-m", help="Model to use")
+    parser.add_argument(
+        "--provider", "-p", choices=sorted(PROVIDER_API_BY_NAME), help="Provider to use"
+    )
+    parser.add_argument("--api-key", "-k", help="API key")
+    parser.add_argument("--base-url", "-u", help="Base URL for API")
+    parser.add_argument(
+        "--openai-compat-auth",
+        choices=("auto", "required", "none"),
+        help="Auth mode for OpenAI-compatible endpoints",
+    )
+    parser.add_argument(
+        "--anthropic-compat-auth",
+        choices=("auto", "required", "none"),
+        help="Auth mode for Anthropic-compatible endpoints",
+    )
+    parser.add_argument(
+        "--insecure-skip-verify",
+        action="store_true",
+        help="Skip TLS verification (e.g. self-signed certs on local providers)",
+    )
+    parser.add_argument(
+        "--continue",
+        "-c",
+        action="store_true",
+        dest="continue_recent",
+        help="Resume the most recent session",
+    )
+    parser.add_argument(
+        "--resume",
+        "-r",
+        dest="resume_session",
+        help="Resume a specific session by ID (full or unique prefix)",
+    )
+    parser.add_argument("--version", action="version", version=f"kon {VERSION}")
+    parser.add_argument(
+        "--extra-tools", help="Comma-separated extra tools to enable (e.g. web_search,web_fetch)"
+    )
+    return parser
+
+
+def main() -> None:
+    args = build_parser().parse_args()
+
+    if args.insecure_skip_verify:
+        config.llm.tls.insecure_skip_verify = True
+
+    extra_tools = (
+        [t.strip() for t in args.extra_tools.split(",") if t.strip()] if args.extra_tools else None
+    )
+
+    from .ui.app import run_tui
+
+    run_tui(args, extra_tools=extra_tools)
+
+
+if __name__ == "__main__":
+    main()
